refactor: convert onwards to library crate
- Create lib.rs to expose public API (AppState, build_router)
- Move tests requiring internal access from main.rs to lib.rs
- Keep main.rs minimal, using library exports
